src/requirements.txt
azure-functions
azure-identity
azure-keyvault-keys
azure-keyvault-secrets
